+++ packages/remix-dev/vite/plugin.ts
@@ -1052,6 +1052,29 @@ export const remixVitePlugin: RemixVitePlugin = (remixUserConfig = {}) => {
             ctx.remixConfig.ssr === false
               ? "spa"
               : "custom",
+
+          ssr: {
+            external: isInRemixMonorepo()
+              ? [
+                  // This is only needed within the Remix repo because these
+                  // packages are linked to a directory outside of node_modules
+                  // so Vite treats them as internal code by default.
+                  "@remix-run/architect",
+                  "@remix-run/cloudflare-pages",
+                  "@remix-run/cloudflare-workers",
+                  "@remix-run/cloudflare",
+                  "@remix-run/css-bundle",
+                  "@remix-run/deno",
+                  "@remix-run/dev",
+                  "@remix-run/express",
+                  "@remix-run/netlify",
+                  "@remix-run/node",
+                  "@remix-run/react",
+                  "@remix-run/serve",
+                  "@remix-run/server-runtime",
+                ]
+              : undefined,
+          },
           optimizeDeps: {
             include: [
               // Pre-bundle React dependencies to avoid React duplicates,
@@ -1527,7 +1550,7 @@ export const remixVitePlugin: RemixVitePlugin = (remixUserConfig = {}) => {
               "",
               `  But other route exports in '${importerShort}' depend on '${id}'.`,
               "",
-              "  See https://remix.run/docs/en/main/future/vite#splitting-up-client-and-server-code",
+              "  See https://remix.run/docs/en/main/guides/vite#splitting-up-client-and-server-code",
               "",
             ].join("\n")
           );
@@ -1539,7 +1562,7 @@ export const remixVitePlugin: RemixVitePlugin = (remixUserConfig = {}) => {
             "",
             `    '${id}' imported by '${importerShort}'`,
             "",
-            "  See https://remix.run/docs/en/main/future/vite#splitting-up-client-and-server-code",
+            "  See https://remix.run/docs/en/main/guides/vite#splitting-up-client-and-server-code",
             "",
           ].join("\n")
         );
@@ -1582,7 +1605,7 @@ export const remixVitePlugin: RemixVitePlugin = (remixUserConfig = {}) => {
             let str = serverOnlyExports.map((e) => `\`${e}\``).join(", ");
             let message =
               `SPA Mode: ${serverOnlyExports.length} invalid route export(s) in ` +
-              `\`${route.file}\`: ${str}. See https://remix.run/future/spa-mode ` +
+              `\`${route.file}\`: ${str}. See https://remix.run/guides/spa-mode ` +
               `for more information.`;
             throw Error(message);
           }
@@ -1595,17 +1618,20 @@ export const remixVitePlugin: RemixVitePlugin = (remixUserConfig = {}) => {
               let message =
                 `SPA Mode: Invalid \`HydrateFallback\` export found in ` +
                 `\`${route.file}\`. \`HydrateFallback\` is only permitted on ` +
-                `the root route in SPA Mode. See https://remix.run/future/spa-mode ` +
+                `the root route in SPA Mode. See https://remix.run/guides/spa-mode ` +
                 `for more information.`;
               throw Error(message);
             }
           }
         }
 
-        return {
-          code: removeExports(code, SERVER_ONLY_ROUTE_EXPORTS),
-          map: null,
-        };
+        let [filepath] = id.split("?");
+
+        return removeExports(code, SERVER_ONLY_ROUTE_EXPORTS, {
+          sourceMaps: true,
+          filename: id,
+          sourceFileName: filepath,
+        });
       },
     },
     {
@@ -1746,6 +1772,12 @@ export const remixVitePlugin: RemixVitePlugin = (remixUserConfig = {}) => {
     },
   ];
 };
+
+function isInRemixMonorepo() {
+  let devPath = path.dirname(require.resolve("@remix-run/dev/package.json"));
+  let devParentDir = path.basename(path.resolve(devPath, ".."));
+  return devParentDir === "packages";
+}
 
 function isEqualJson(v1: unknown, v2: unknown) {
   return JSON.stringify(v1) === JSON.stringify(v2);