+++ templates/classic-remix-compiler/arc/server.ts
@@ -4,6 +4,6 @@ import { installGlobals } from "@remix-run/node";
 import sourceMapSupport from "source-map-support";
 
 sourceMapSupport.install();
-installGlobals();
+installGlobals({ nativeFetch: true });
 
 export const handler = createRequestHandler({ build });