+++ packages/remix-cloudflare-pages/CHANGELOG.md
@@ -1,5 +1,19 @@
 # `@remix-run/cloudflare-pages`
 
+## 2.9.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/cloudflare@2.9.1`
+
+## 2.9.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/cloudflare@2.9.0`
+
 ## 2.8.1
 
 ### Patch Changes